[#64] remove old components
