+++ app/components/places/detail/place-detail.html
@@ -3,11 +3,11 @@
 <GridLayout rows="*,auto">
   <StackLayout row="0">
     <ScrollView orientation="vertical" *ngIf="isReady">
-      <kl-form-place-edit [place]="place"></kl-form-place-edit>
+      <kl-form-place-validation [place]="place" (submit)="onSubmit($event)"></kl-form-place-validation>
     </ScrollView>
   </StackLayout>
   <StackLayout row="0" verticalAlignment="center" *ngIf="!isReady">
     <ActivityIndicator busy="true"></ActivityIndicator>
   </StackLayout>
-  <kl-places-action-bar row="1"></kl-places-action-bar>
+  <StackLayout row="1" kl-places-action-bar></StackLayout>
 </GridLayout>
