+++ app/components/places/pin/pin-place.html
@@ -1,8 +1,11 @@
 <ActionBar title="Pin places" class="action-bar">
-  <ActionItem [ios.systemIcon]="searchView === 'map' ? 12 : 11" ios.position="right" (tap)="onToggleSearchView()"></ActionItem>
+  <ActionItem [ios.systemIcon]="11" ios.position="right" (tap)="onToggleSearchView()"></ActionItem>
 </ActionBar>
-<GridLayout rows="*,auto" backgroundColor="aqua" height="100%">
-  <StackLayout height="100%" orientation="vertical" row="0" *ngIf="step === 'find'" kl-place-pin-search></StackLayout>
+<GridLayout rows="*,auto" height="100%">
+  <StackLayout height="100%" orientation="vertical" row="0">
+    <kl-place-pin-search #search (findPlace)="onFindPlace($event)" *ngIf="step === 'find'"></kl-place-pin-search>
+    <kl-form-place-validation *ngIf="step === 'validation'" [place]="place" (submit)="onSubmit($event)"></kl-form-place-validation>
+  </StackLayout>
   <StackLayout row="1" activePage="pin" kl-places-action-bar></StackLayout>
 </GridLayout>
