--- app/modules/form-place/components/add/form-place-add.html
@@ -1,5 +1,0 @@
-<StackLayout orientation="vertical" height="100%">
-  <kl-form-place-map (submit)="onFindPlace($event)" *ngIf="step === 'search' && searchView === 'map'"></kl-form-place-map>
-  <kl-form-place-search (submit)="onFindPlace($event)" *ngIf="step === 'search' && searchView === 'name'"></kl-form-place-search>
-  <kl-form-place-validation *ngIf="step === 'validation'" [place]="place" (submit)="onSubmit($event)"></kl-form-place-validation>
-</StackLayout>
