--- app/modules/form-place/components/map/form-place-map.html
@@ -1,25 +1,0 @@
-<GridLayout rows="*,*" backgroundColor="aqua" height="100%">
-  <StackLayout row="0" verticalAlignment="top" height="100%">
-
-    <MapView #mapView height="100%"
-             [zoom]="zoom" [bearing]="bearing"
-             [tilt]="tilt" i-padding="50,50,50,50" [padding]="padding" (mapReady)="onMapReady($event)"
-             (markerSelect)="onMarkerEvent($event)" (markerBeginDragging)="onMarkerEvent($event)"
-             (markerEndDragging)="onMarkerEvent($event)" (markerDrag)="onMarkerEvent($event)"
-             (markerInfoWindowTapped)="onMarkerEvent($event)" (coordinateTapped)="onCoordinateTapped($event)"
-             (cameraChanged)="onCameraChanged($event)"></MapView>
-
-
-  </StackLayout>
-  <StackLayout row="1" height="100%" id="tom">
-    <ListView [items]="places" separatorColor="transparent" class="list-group" height="100%">
-      <ng-template let-item="item" let-i="index" let-odd="odd" let-even="even">
-        <FlexboxLayout backgroundColor="#f6f8fa">
-          <Image src="{{item.type | placeIconImgSrc}}" stretch="none" horizontalAlignment="center"></Image>
-          <Label [text]="item.name" marginLeft="10px" flexGrow="1" (tap)="onItemNameTap(item)"></Label>
-          <Label text="{{ iconAdd }}" class="fa fa-2" color="#1b926c" paddingLeft="10" paddingRight="10" (tap)="onItemAdd(item)"></Label>
-        </FlexboxLayout>
-      </ng-template>
-    </ListView>
-  </StackLayout>
-</GridLayout>
