--- app/modules/form-place/components/search/form-place-search.html
@@ -1,4 +1,0 @@
-<StackLayout orientation="vertical">
-  <kl-search-place (selected)="onPlaceSelected($event)"></kl-search-place>
-</StackLayout>
-
