+++ app/modules/form-place/components/validation/form-place-validation.common.css
@@ -31,7 +31,7 @@
 }
 
 .place-comments-container TextView {
-    height: 200;
+    height: 150;
 }
 
 .place-submit-container {
