--- app/modules/form-place/form-place.routes.ts
@@ -1,12 +1,0 @@
-import { FormPlaceSearchComponent } from './components/search/form-place-search.component';
-import { FormPlaceValidationComponent } from './components/validation/form-place-validation.component';
-
-export const routes = [
-  { path: 'add-place', component: FormPlaceSearchComponent },
-  { path: 'add-place-validation/:id/:origin', component: FormPlaceValidationComponent }
-]
-export const navigatableComponents = [
-  FormPlaceSearchComponent,
-  FormPlaceValidationComponent
-];
-
