+++ app/modules/google-sdk/components/map/map.html
@@ -1,4 +1,4 @@
-<MapView height="100%"
+<MapView #mapView height="100%"
              [zoom]="zoom" [bearing]="bearing"
              [tilt]="tilt" i-padding="50,50,50,50" [padding]="padding" (mapReady)="onMapReady($event)"
              (markerSelect)="onMarkerEvent($event)" (markerBeginDragging)="onMarkerEvent($event)"
